Move server nav components into a dedicated folder
